Added delete functionality and also update functionality

+++ routes/ideas.js
@@ -45,18 +45,26 @@ try {
 
 router.put('/:id', async (req, res) => {
     try {
-        const updatedIdea = await Idea.findByIdAndUpdate(
-            req.params.id,
-                {   $set: {
-                            text: req.body.text,
-                            tag: req.body.tag
-                         }
-                },
-                {
-                    new: true
-                }               
-        );
-        res.json({ success: true, data: updatedIdea});
+
+        const idea = await Idea.findById(req.params.id);
+        if (idea.username === req.body.username) {
+            const updatedIdea = await Idea.findByIdAndUpdate(
+                req.params.id,
+                    {   $set: {
+                                text: req.body.text,
+                                tag: req.body.tag
+                             }
+                    },
+                    {
+                        new: true
+                    }               
+            );
+            return res.json({ success: true, data: updatedIdea});
+        }
+
+        // Username does not match
+        res.status(403).json({success: false, error: 'You are not authorised to update this resource.'});
+
     } catch (error) {
         console.log(error);
         res.status(500).json({success: false, error: 'Something went wrong'});
@@ -69,8 +77,17 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
     try {
-        await Idea.findByIdAndDelete(req.params.id);
-        res.json({ success: true, data: {}});
+        const idea = await Idea.findById(req.params.id);
+
+        // Match the usernames
+        if (idea.username === req.body.username) {
+            await Idea.findByIdAndDelete(req.params.id);
+            return res.json({ success: true, data: {}});
+        }
+
+        //Usernames do  not match
+        res.status(403).json({success: false, error: 'You are not authorised to delete this resource.'});
+       
     } catch (error) {
         console.log(error);
         res.status(500).json({success: false, error: 'Something went wrong'});
